Adding the ability to share and print
