+++ public/global.css
@@ -94,6 +94,22 @@ button:hover, .btn:hover {
   background: rgb(0, 105, 217);
 }
 
+.svg-button {
+  background: transparent;
+  color: inherit;
+  box-shadow: none;
+  border: none;
+  padding: 0;
+}
+.svg-button:hover {
+  background: transparent;
+}
+.svg-button svg {
+  color: #1B3A52;
+  height: 20px;
+  width: 20px;
+}
+
 hr {
   margin: 1.5rem 0;
   border: 0;
