+++ src/routes/Configuration.svelte
@@ -19,6 +19,7 @@
     contactEmail,
     contactAffiliation
   } from '../settingsStore';
+  import { shareDashboard } from '../util/share';
 
   export let pin;
   export let productUID;
@@ -114,32 +115,6 @@
     }
   }
 
-  const shareDevice = () => {
-    if (navigator.share) { // Share Device URL with Web Share API
-      navigator.share({
-        title: 'Device Dashboard',
-        url: `https://airnote.live/${deviceUID}`
-      }).then(() => {
-        notifier.success('Thanks for sharing!');
-      })
-      .catch(console.error);
-    } else {
-      // Fallback to copying text to the users clipboard with a
-      // hacky cross-browser approach.
-      const el = document.createElement('textarea');
-      el.value = `https://airnote.live/${deviceUID}`;
-      el.setAttribute('readonly', '');
-      el.style.position = 'absolute';
-      el.style.left = '-9999px';
-      document.body.appendChild(el);
-      el.select();
-      document.execCommand('copy');
-      document.body.removeChild(el);
-
-      notifier.success('Dashboard URL Copied to clipboard!');
-    }
-  }
-
   onMount(async () => {
     // If no Pin, Get Env Vars using legacy req API
     if (pin === '') {
@@ -228,7 +203,7 @@
   </Col>
 </div>
 <div class="share">
-  <button on:click={shareDevice}>Share my Device Dashboard</button>
+  <button on:click={() => shareDashboard(deviceUID)}>Share my Device Dashboard</button>
 </div>
 <hr />
 <p>
